+++ Tema_5/Ejmplo/Amazon/Amazon_Montecastelo/Database/Models/DetalleVenta.cs
@@ -16,8 +16,8 @@ namespace Amazon_Montecastelo.Database.Models
         public Productos Producto { get; set; }
 
         public int Cantidad { get; set; }
-        public decimal PrecioUnitario { get; set; }
-        public decimal PrecioTotal { get; set; }
+        public double PrecioUnitario { get; set; }
+        public double PrecioTotal { get; set; }
